adjust shader source header file
